wip seperating out blmm_inference
